test persimon in smaller size

+++ tests/models/persimmon/test_modeling_persimmon.py
@@ -385,11 +385,11 @@ class PersimmonIntegrationTest(unittest.TestCase):
     @slow
     def test_model_8b_chat_logits(self):
         input_ids = [1, 306, 4658, 278, 6593, 310, 2834, 338]
-        model = PersimmonForCausalLM.from_pretrained("ArthurZ/persimmon-8b-chat", device_map="auto")
+        model = PersimmonForCausalLM.from_pretrained("ArthurZ/persimmon-8b-chat", device_map="auto", torch_dtype = torch.float16)
         out = model(torch.tensor([input_ids])).logits
 
         EXPECTED_MEAN = torch.tensor(
-            [[-11.2879, -11.2628, -11.2498, -11.2534, -11.2676, -11.2638, -11.2501, -11.2431]], dtype=torch.float32
+            [[-11.2879, -11.2628, -11.2498, -11.2534, -11.2676, -11.2638, -11.2501, -11.2431]], dtype=torch.float16
         )
         torch.testing.assert_close(out.cpu().mean(-1), EXPECTED_MEAN, atol=1e-4, rtol=1e-4)
         # fmt: off
@@ -404,7 +404,7 @@ class PersimmonIntegrationTest(unittest.TestCase):
         prompt = "human: Simply put, the theory of relativity states that?\n\nadept:"
         tokenizer = AutoTokenizer.from_pretrained("ArthurZ/persimmon-8b-chat", use_fast=False)
         input_ids = tokenizer.encode(prompt, return_tensors="pt").to(torch_device)
-        model = PersimmonForCausalLM.from_pretrained("ArthurZ/persimmon-8b-chat").to(torch_device)
+        model = PersimmonForCausalLM.from_pretrained("ArthurZ/persimmon-8b-chat", torch_dtype = torch.float16).to(torch_device)
 
         # greedy generation outputs
         generated_ids = model.generate(input_ids, max_new_tokens=64)
